Restructure packages, add Nomad and Consul clients to Config struct to re-use rather than re-creating and re-connecting to services during each tick.
